Update alpha_to_html-lpac.pl
table conversion ready.  converts each tab or run of 3 or more spaces to a column boundary.  Converts each line break as a row boundary.  
'<table'  is converted to the beginning of the table, 'table>' to the end of the table. Table rendered as divs with classes 'table','tr','th' and 'td'